+++ tests/test_config_parser.py
@@ -128,7 +128,7 @@ class MkosiConfig:
             "qemu_headless": False,
             "qemu_smp": "2",
             "qemu_mem": "1G",
-            "network_veth": False,
+            "netdev": False,
             "ephemeral": False,
             "with_unified_kernel_images": True,
             "hostonly_initrd": False,
@@ -366,8 +366,8 @@ class MkosiConfig:
                 self._append_list("extra_search_paths", mk_config_host["ExtraSearchPaths"], job_name, ":")
             if "QemuHeadless" in mk_config_host:
                 self.reference_config[job_name]["qemu_headless"] = mk_config_host["QemuHeadless"]
-            if "NetworkVeth" in mk_config_host:
-                self.reference_config[job_name]["network_veth"] = mk_config_host["NetworkVeth"]
+            if "Netdev" in mk_config_host:
+                self.reference_config[job_name]["netdev"] = mk_config_host["Netdev"]
             if "Ephemeral" in mk_config_host:
                 self.reference_config[job_name]["ephemeral"] = mk_config_host["Ephemeral"]
             if "Ssh" in mk_config_host:
@@ -586,7 +586,7 @@ class MkosiConfigManyParams(MkosiConfigOne):
             "Host": {
                 "ExtraSearchPaths": "search/here:search/there",
                 "QemuHeadless": True,
-                "NetworkVeth": True,
+                "Netdev": True,
             },
         }
         self._prepare_mkosi_default(directory, mk_config)
@@ -652,7 +652,7 @@ class MkosiConfigManyParams(MkosiConfigOne):
             "Host": {
                 "ExtraSearchPaths": "search/ubu",
                 "QemuHeadless": True,
-                "NetworkVeth": True,
+                "Netdev": True,
             },
         }
         self._prepare_mkosi_default_d(directory, mk_config, 1)
@@ -718,7 +718,7 @@ class MkosiConfigManyParams(MkosiConfigOne):
             "Host": {
                 "ExtraSearchPaths": "search/debi",
                 "QemuHeadless": True,
-                "NetworkVeth": True,
+                "Netdev": True,
             },
         }
         self._prepare_mkosi_default_d(directory, mk_config, 2)